+++ workers/resolver.php
@@ -1,6 +1,6 @@
 <?php
 
-require __DIR__ . '/vendor/autoload.php';
+require __DIR__ . '/../vendor/autoload.php';
 
 $bootstrap = new Movim\Bootstrap;
 $bootstrap->boot(true);